moved RecordedResponse into separate file
