+++ Sources/Networking/ElloProvider_Specs.swift
@@ -4,22 +4,6 @@
 
 import Moya
 
-public struct RecordedResponse {
-    let endpoint: ElloAPI
-    let responseClosure: (target: ElloAPI) -> EndpointSampleResponse
-
-    public init(endpoint: ElloAPI, responseClosure: (target: ElloAPI) -> EndpointSampleResponse) {
-        self.endpoint = endpoint
-        self.responseClosure = responseClosure
-    }
-
-    public init(endpoint: ElloAPI, response: EndpointSampleResponse) {
-        self.endpoint = endpoint
-        self.responseClosure = { _ in return response }
-    }
-
-}
-
 
 public struct ElloProvider_Specs {
     public static var errorStatusCode: ErrorStatusCode = .Status404
